move plutovista code to extension
